refactor(omp): group provider subsystem
