+++ src/agent-cli-provider/index.ts
@@ -40,12 +40,12 @@ export {
   type OmpRpcTaskHooks,
   type OmpRpcTaskRequest,
   type OmpRpcTaskResult,
-} from './omp-rpc-driver';
+} from './omp/rpc-driver';
 export { buildOmpPrompt } from './adapters/omp';
-export { DEFAULT_OMP_RPC_DECODER_LIMITS } from './omp-rpc-protocol';
-export { ABORT_GRACE_MS, EXIT_GRACE_MS } from './omp-rpc-bounds';
-export { OMP_SUPPORTED_VERSION } from './omp-release';
-export type { OmpSessionLaunch, VerifiedOmpPartition, VerifiedOmpSessionFile } from './omp-rpc-session';
+export { DEFAULT_OMP_RPC_DECODER_LIMITS } from './omp/rpc-protocol';
+export { ABORT_GRACE_MS, EXIT_GRACE_MS } from './omp/rpc-bounds';
+export { OMP_SUPPORTED_VERSION } from './omp/release';
+export type { OmpSessionLaunch, VerifiedOmpPartition, VerifiedOmpSessionFile } from './omp/rpc-session';
 export {
   OmpSdkProcessRunnerError,
   runOmpSdkProcess,
@@ -53,7 +53,7 @@ export {
   type OmpSdkProcessResult,
   type OmpSdkProcessRunnerOptions,
   type OmpSdkRunningProcess,
-} from './omp-sdk-process-runner';
+} from './omp/sdk-process-runner';
 export {
   detectRuntimeProviderCliFeatures,
   prepareSingleAgentProviderCommand,
@@ -114,7 +114,7 @@ export {
   type OmpSdkSidecarRequest,
   type OmpSdkTextSidecarRequest,
   type OmpSdkToolId,
-} from './omp-sdk-protocol';
+} from './omp/sdk-protocol';
 
 
 export type {
